refactor(quality): split SigningService below the class-length threshold
Extract SignedArtifactProducer (produce + store the verifiable signed
artifact, resolve the document node and signer label) and
SigningRequestValidator (request data + provider/level honesty at creation).
Both moved verbatim; SigningService keeps the lifecycle orchestration and the
conclusion emitter. Applied phpcbf to the touched files.

phpmd is now 0 findings (exit 0) for docudesk.
